feat(python): add Python bindings with PyO3 and maturin build system
- Add PyO3 bindings for Rust CLI functionality
- Configure maturin for Python package building
- Add Python 3.13 support with uv package management
- Add MIT license and pyproject.toml configuration
- Update Makefile.toml with Python build tasks
- Add .python-version for consistent Python version
